# src/main/resources/api/api-spec.yaml
openapi: 3.0.3
info:
  title: AURA Client API
  description: API for client management
  version: 1.0.0
paths:
  /clients:
    get:
      summary: List all clients
      operationId: listClients
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      summary: Create a new client
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
  /clients/{clientId}:
    get:
      summary: Get a client by ID
      operationId: getClientById
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
components:
  schemas:
    Address:
      type: object
      properties:
        street:
          type: string
        number:
          type: string
        zipcode:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
      required:
        - street
        - number
        - zipcode
        - neighborhood
        - city
    ClientRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
      required:
        - firstName
        - lastName
        - cpf
        - birthDate
        - addresses
    Client:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'